# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Sales Platform API
  version: 2.0.0
  description: Authenticated, tenant-isolated CRM, reporting, settings, and AI workflows.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [CRM]
      summary: List a cursor-paginated tenant lead page
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string, maxLength: 200 } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/LeadStatus" } }
      responses:
        "200": { description: Lead page }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      operationId: createLead
      tags: [CRM]
      summary: Create a tenant-owned lead
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadInput" }
      responses:
        "201": { description: Lead created }
        "400": { $ref: "#/components/responses/BadRequest" }
  /leads/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    put:
      operationId: updateLead
      tags: [CRM]
      summary: Update a tenant-owned lead
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadInput" }
      responses:
        "200": { description: Lead updated }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteLead
      tags: [CRM]
      summary: Delete a tenant-owned lead
      responses:
        "204": { description: Lead deleted }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [Campaigns]
      summary: List owned campaigns
      responses:
        "200": { description: Campaigns returned }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createCampaign
      tags: [Campaigns]
      summary: Create an unsent draft campaign
      responses:
        "201": { description: Draft campaign created }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /unsubscribe:
    get:
      operationId: showUnsubscribeConfirmation
      tags: [Campaigns]
      security: []
      summary: Show confirmation for a signed tenant-scoped unsubscribe capability
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string, minLength: 20, maxLength: 1000 }
      responses:
        "200": { description: Confirmation page }
        "400": { description: Invalid signed capability }
    post:
      operationId: oneClickUnsubscribe
      tags: [Campaigns]
      security: []
      summary: Suppress a recipient and cancel pending messages using a signed one-click capability
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string, minLength: 20, maxLength: 1000 }
      responses:
        "204": { description: Idempotent non-enumerating unsubscribe response }
        "429": { description: Rate limit reached }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: refresh_token }
  schemas:
...
    LeadStatus:
      type: string
      enum: [INTERESTED, MEETING, FOLLOW_UP, PROPOSAL_SENT, CLOSED, LOST]
    LeadInput:
      type: object
      properties:
        company: { type: string, maxLength: 160 }
        contact: { type: string, maxLength: 160 }
        email: { type: [string, "null"], format: email, maxLength: 254 }
        phone: { type: [string, "null"], maxLength: 50 }
        industry: { type: [string, "null"], maxLength: 120 }
        status: { $ref: "#/components/schemas/LeadStatus" }
        value: { type: number, minimum: 0, maximum: 9999999999999.99 }
        notes: { type: [string, "null"], maxLength: 10000 }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
...
  responses:
    BadRequest: { description: Validation or token failure, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Unauthorized: { description: Authentication failed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
    NotFound: { description: Resource missing or belongs to another tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...